Separate impl from tests
